pyproject.toml
[tool.poetry]
name = "tovar-zapolnenie"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fast-bitrix24 = "^1.7.3"
loguru = "^0.7.2"
flask = "^3.0.3"
flask-restx = "^1.3.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
